tweak: complete suggestions and remove the need to convert to record

+++ products/src/main/java/com/alexx666/products/infra/InMemoryProductDAO.java
@@ -89,7 +89,7 @@ public class InMemoryProductDAO implements ProductsDAO {
             return 0;
         }
 
-        int totalRating = 0;
+        double totalRating = 0;
 
         for (Integer rating: productRatings.values()) {
             totalRating += rating;
